Refs #16986 - Improve style, fix comments
