+++ webpack/components/NewTemplateSync/components/NewTemplateSyncForm/NewTemplateSyncFormConstants.js
@@ -1,1 +1,1 @@
-export const formName = 'newTemplateSync';
+export const NEW_TEMPLATE_SYNC_FORM_NAME = 'newTemplateSync';
