+++ webpack/components/NewTemplateSync/components/TextButtonField/CheckboxField.js
@@ -3,7 +3,7 @@ import { Checkbox } from 'patternfly-react';
 import PropTypes from 'prop-types';
 
 const CheckboxField = ({ input, item, disabled }) => (
-  <Checkbox {...input} disabled={disabled} />
+  <Checkbox {...input} disabled={disabled} checked={input.value} />
 );
 
 CheckboxField.propTypes = {
